+++ format/runlength.go
@@ -26,7 +26,11 @@ import (
 //
 //  - "stream.Broadcast":
 //    Formatter: "format.Runlength"
+//    RunlengthSeparator: ":"
 //    RunlengthFormatter: "format.Envelope"
+//
+// RunlengthSeparator sets the separator character placed after the runlength.
+// This is set to ":" by default.
 //
 // RunlengthDataFormatter defines the formatter for the data transferred as
 // message. By default this is set to "format.Forward"