+++ format/runlength_test.go
@@ -1,0 +1,37 @@
+// Copyright 2015-2016 trivago GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package format
+
+import (
+	"github.com/trivago/gollum/core"
+	"github.com/trivago/tgo/ttesting"
+	"testing"
+)
+
+func TestRunlength(t *testing.T) {
+	expect := ttesting.NewExpect(t)
+
+	config := core.NewPluginConfig("", "format.Runlength")
+	plugin, err := core.NewPlugin(config)
+	expect.NoError(err)
+
+	formatter, casted := plugin.(*Runlength)
+	expect.True(casted)
+
+	msg := core.NewMessage(nil, []byte("test"), 0)
+	formatter.Format(msg)
+
+	expect.Equal("4:test", string(msg.Data))
+}