+++ format/splittojson_test.go
@@ -1,0 +1,98 @@
+// Copyright 2015-2016 trivago GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package format
+
+import (
+	"encoding/json"
+	"github.com/trivago/gollum/core"
+	"github.com/trivago/tgo/tcontainer"
+	"github.com/trivago/tgo/ttesting"
+	"testing"
+)
+
+func TestSplitToJSON(t *testing.T) {
+	expect := ttesting.NewExpect(t)
+
+	config := core.NewPluginConfig("", "format.SplitToJSON")
+	config.Override("SplitToJSONToken", ",")
+	config.Override("SplitToJSONKeys", []string{"first", "second", "third"})
+
+	plugin, err := core.NewPlugin(config)
+	expect.NoError(err)
+
+	formatter, casted := plugin.(*SplitToJSON)
+	expect.True(casted)
+
+	msg := core.NewMessage(nil, []byte("test1,test2,test3"), 10)
+	formatter.Format(msg)
+
+	jsonData := tcontainer.NewMarshalMap()
+	err = json.Unmarshal(msg.Data, &jsonData)
+	expect.NoError(err)
+
+	expect.MapEqual(jsonData, "first", "test1")
+	expect.MapEqual(jsonData, "second", "test2")
+	expect.MapEqual(jsonData, "third", "test3")
+}
+
+func TestSplitToJSONTooFew(t *testing.T) {
+	expect := ttesting.NewExpect(t)
+
+	config := core.NewPluginConfig("", "format.SplitToJSON")
+	config.Override("SplitToJSONToken", ",")
+	config.Override("SplitToJSONKeys", []string{"first", "second"})
+
+	plugin, err := core.NewPlugin(config)
+	expect.NoError(err)
+
+	formatter, casted := plugin.(*SplitToJSON)
+	expect.True(casted)
+
+	msg := core.NewMessage(nil, []byte("test1,test2,test3"), 10)
+	formatter.Format(msg)
+
+	jsonData := tcontainer.NewMarshalMap()
+	err = json.Unmarshal(msg.Data, &jsonData)
+	expect.NoError(err)
+
+	expect.MapEqual(jsonData, "first", "test1")
+	expect.MapEqual(jsonData, "second", "test2")
+	expect.MapNotSet(jsonData, "third")
+}
+
+func TestSplitToJSONTooMany(t *testing.T) {
+	expect := ttesting.NewExpect(t)
+
+	config := core.NewPluginConfig("", "format.SplitToJSON")
+	config.Override("SplitToJSONToken", ",")
+	config.Override("SplitToJSONKeys", []string{"first", "second", "third", "fourth"})
+
+	plugin, err := core.NewPlugin(config)
+	expect.NoError(err)
+
+	formatter, casted := plugin.(*SplitToJSON)
+	expect.True(casted)
+
+	msg := core.NewMessage(nil, []byte("test1,test2,test3"), 10)
+	formatter.Format(msg)
+
+	jsonData := tcontainer.NewMarshalMap()
+	err = json.Unmarshal(msg.Data, &jsonData)
+	expect.NoError(err)
+
+	expect.MapEqual(jsonData, "first", "test1")
+	expect.MapEqual(jsonData, "second", "test2")
+	expect.MapEqual(jsonData, "third", "test3")
+}